+++ pipelines/asana_dlt/__init__.py
@@ -2,19 +2,21 @@
 import typing as t
 from typing import Sequence
 import dlt
-from asana import Client as AsanaClient
 
 from dlt.extract.source import DltResource
-
-DEFAULT_START_DATE = "2010-01-01T00:00:00.000Z"
-
-
-def get_client(
-    access_token: str,
-) -> AsanaClient:
-    """Returns an Asana client with a valid access token"""
-    asana = AsanaClient.access_token(access_token)
-    return asana
+from .const import (
+    PROJECT_FIELDS,
+    USER_FIELDS,
+    DEFAULT_START_DATE,
+    TIMEOUT,
+    SECTION_FIELDS,
+    TAG_FIELDS,
+    TASK_FIELDS,
+    STORY_FIELDS,
+    TEAMS_FIELD,
+    WORKSPACE_FIELDS,
+)
+from .helpers import get_client
 
 
 @dlt.resource(write_disposition="replace")
@@ -22,9 +24,7 @@ def workspaces(access_token: str = dlt.secrets.value) -> t.Iterator[dict]:
     """Returns a list of workspaces"""
     print("Fetching workspaces...")
     yield from get_client(access_token).workspaces.find_all(
-        opt_fields=",".join(
-            ["gid", "name", "is_organization", "resource_type", "email_domains"]
-        )
+        opt_fields=",".join(WORKSPACE_FIELDS)
     )
     print("Done fetching workspaces.")
 
@@ -41,41 +41,8 @@ def projects(workspace, access_token: str = dlt.secrets.value) -> t.Iterator[dic
         return list(
             get_client(access_token).projects.find_all(
                 workspace=workspace["gid"],
-                timeout=300,
-                opt_fields=",".join(
-                    [
-                        "name",
-                        "gid",
-                        "owner",
-                        "current_status",
-                        "custom_fields",
-                        "default_view",
-                        "due_date",
-                        "due_on",
-                        "is_template",
-                        "created_at",
-                        "modified_at",
-                        "start_on",
-                        "archived",
-                        "public",
-                        "members",
-                        "followers",
-                        "color",
-                        "notes",
-                        "icon",
-                        "permalink_url",
-                        "workspace",
-                        "team",
-                        "resource_type",
-                        "current_status_update",
-                        "custom_field_settings",
-                        "completed",
-                        "completed_at",
-                        "completed_by",
-                        "created_from_template",
-                        "project_brief",
-                    ]
-                ),
+                timeout=TIMEOUT,
+                opt_fields=",".join(PROJECT_FIELDS),
             )
         )
     finally:
@@ -96,17 +63,8 @@ def sections(project_array, access_token: str = dlt.secrets.value) -> t.Iterator
             for project in project_array
             for section in get_client(access_token).sections.get_sections_for_project(
                 project_gid=project["gid"],
-                timeout=300,
-                opt_fields=",".join(
-                    [
-                        "gid",
-                        "resource_type",
-                        "name",
-                        "created_at",
-                        "project",
-                        "projects",
-                    ]
-                ),
+                timeout=TIMEOUT,
+                opt_fields=",".join(SECTION_FIELDS),
             )
         ]
     finally:
@@ -123,20 +81,8 @@ def tags(workspace, access_token: str = dlt.secrets.value) -> t.Iterator[dict]:
             tag
             for tag in get_client(access_token).tags.find_all(
                 workspace=workspace["gid"],
-                timeout=300,
-                opt_fields=",".join(
-                    [
-                        "gid",
-                        "resource_type",
-                        "created_at",
-                        "followers",
-                        "name",
-                        "color",
-                        "notes",
-                        "permalink_url",
-                        "workspace",
-                    ]
-                ),
+                timeout=TIMEOUT,
+                opt_fields=",".join(TAG_FIELDS),
             )
         ]
     finally:
@@ -158,43 +104,9 @@ def tasks(
         for project in project_array
         for task in get_client(access_token).tasks.find_all(
             project=project["gid"],
-            timeout=300,
+            timeout=TIMEOUT,
             modified_since=modified_at.start_value,
-            opt_fields=",".join(
-                [
-                    "gid",
-                    "resource_type",
-                    "name",
-                    "approval_status",
-                    "assignee_status",
-                    "created_at",
-                    "assignee",
-                    "start_on",
-                    "start_at",
-                    "due_on",
-                    "due_at",
-                    "completed",
-                    "completed_at",
-                    "completed_by",
-                    "modified_at",
-                    "dependencies",
-                    "dependents",
-                    "external",
-                    "notes",
-                    "num_subtasks",
-                    "resource_subtype",
-                    "followers",
-                    "parent",
-                    "permalink_url",
-                    "tags",
-                    "workspace",
-                    "custom_fields",
-                    "project",
-                    "memberships",
-                    "memberships.project.name",
-                    "memberships.section.name",
-                ]
-            ),
+            opt_fields=",".join(TASK_FIELDS),
         )
     )
     print(f"Done fetching tasks for {len(project_array)} projects.")
@@ -213,35 +125,8 @@ def stories(task, access_token: str = dlt.secrets.value) -> t.Iterator[dict]:
             story
             for story in get_client(access_token).stories.get_stories_for_task(
                 task_gid=task["gid"],
-                timeout=300,
-                opt_fields=",".join(
-                    [
-                        "gid",
-                        "resource_type",
-                        "created_at",
-                        "created_by",
-                        "resource_subtype",
-                        "text",
-                        "is_pinned",
-                        "assignee",
-                        "dependency",
-                        "follower",
-                        "new_section",
-                        "old_section",
-                        "new_text_value",
-                        "old_text_value",
-                        "preview",
-                        "project",
-                        "source",
-                        "story",
-                        "tag",
-                        "target",
-                        "task",
-                        "sticker_name",
-                        "custom_field",
-                        "type",
-                    ]
-                ),
+                timeout=TIMEOUT,
+                opt_fields=",".join(STORY_FIELDS),
             )
         ]
     finally:
@@ -261,18 +146,8 @@ def teams(workspace, access_token: str = dlt.secrets.value) -> t.Iterator[dict]:
             team
             for team in get_client(access_token).teams.find_by_organization(
                 organization=workspace["gid"],
-                timeout=300,
-                opt_fields=",".join(
-                    [
-                        "gid",
-                        "resource_type",
-                        "name",
-                        "description",
-                        "organization",
-                        "permalink_url",
-                        "visibility",
-                    ]
-                ),
+                timeout=TIMEOUT,
+                opt_fields=",".join(TEAMS_FIELD),
             )
         ]
     finally:
@@ -292,10 +167,8 @@ def users(workspace, access_token: str = dlt.secrets.value) -> t.Iterator[dict]:
             user
             for user in get_client(access_token).users.find_all(
                 workspace=workspace["gid"],
-                timeout=300,
-                opt_fields=",".join(
-                    ["gid", "resource_type", "name", "email", "photo", "workspaces"]
-                ),
+                timeout=TIMEOUT,
+                opt_fields=",".join(USER_FIELDS),
             )
         ]
     finally:
